refactor: 게시물 api 분리
